refactor(cli): rename oclif dispatch module
